Add very-soft-if-deps
This probably goes against everything that taskgraph should be used for
but it works...

This is a workaround for https://github.com/taskcluster/taskgraph/issues/710

Basically I create the task without deps but an extra attribute that I
transform into dependencies at morph time once the rest of the graph is
decided. Which means that we're potentially removing the task at morph
time if there's no dependency that would bring it in, and I'm very much
not handling this properly because I don't have the use for it.
Basically this will only work if the task is at the very end of a graph
branch, which is the case for my webhook task.